requirements.txt
streamlit>=1.25.0
numpy>=1.21.0
scikit-learn>=1.1.0
pillow>=9.0.0
